#11: Move dummy raindrop fixture to conftest
